+++ lib/tarzan.rb
@@ -1,3 +1,1 @@
-require 'tarzan/interfaces/shell/interface'
-require 'tarzan/interfaces/campfire/interface'
-require 'tarzan/interfaces/desktop/interface'
+require 'tarzan/tournament'