+++ lib/tarzan/games/rock_paper_scissors/bot.rb
@@ -1,4 +1,0 @@
-#http://www.nytimes.com/interactive/science/rock-paper-scissors.html?_r=0
-# RRRS SPRR > R
-# SRRR PSPR > P
-# PSRR SPSP > R