+++ lib/tarzan/interface.rb
@@ -1,0 +1,13 @@
+module Tarzan
+  module Interface
+    def self.included(base)
+      base.extend base
+    end
+
+    def run(options = {})
+    end
+
+    def enqueue(message, choices = {})
+    end
+  end
+end