+++ lib/tarzan/players/random.rb
@@ -1,0 +1,1 @@
+# This will always make a random choice