+++ lib/tarzan/players/smart.rb
@@ -1,0 +1,6 @@
+# This will always make a smart choice
+
+# For instance see http://www.nytimes.com/interactive/science/rock-paper-scissors.html?_r=0
+# RRRS SPRR > R
+# SRRR PSPR > P
+# PSRR SPSP > R