+++ HISTORY.md
@@ -1,3 +1,9 @@
+v0.0.4  - 2014/02/09
+
+* Rewritten with Module, no Singleton class
+* More compact files for each game and interface
+* Room for separate player modules (Human, Random, Smart)
+
 v0.0.3  - 2014/02/05
 
 * Add Odd Or Evens game